Implementação CarouselView
- Inicio dos testes com Snapshot
- Atualização de métodos para retornar as imagens via ViewModel e Delegate (em andamento).
