+++ solutions/devsprint-viewcode-1/ViewCodeChallenge-RealEstate/Screens/Components/CarouselView/Cells/CarouselCellViewModel.swift
@@ -1,0 +1,17 @@
+//
+//  CarouselCellViewModel.swift
+//  ViewCodeChallenge-RealEstate
+//
+//  Created by Elena Diniz on 21/10/21.
+//
+
+import Foundation
+
+struct CarouselCellViewModel {
+    
+    let images: Property
+    
+    init(images: Property) {
+        self.images = images
+    }
+}
