+++ solutions/devsprint-viewcode-1/ViewCodeChallenge-RealEstateTests/CarouselViewTest.swift
@@ -15,7 +15,7 @@ class CarouselViewTest: XCTestCase {
         
         let carouselView = CarouselView()
         
-        assertSnapshot(matching: carouselView, as: .image, record: true)
+        assertSnapshot(matching: carouselView, as: Snapshotting.image(size: CGSize(width: 414, height: 200)), record: true)
         
         
     }
